♻️️ Trekk funskjon for å opprette testdata for vilkårvurderinger ut i egen utils
